--- src/main/java/com/lhs/chatting/util/RoomNumberGenerateUtils.java
@@ -1,13 +1,0 @@
-package com.lhs.chatting.util;
-
-import lombok.experimental.UtilityClass;
-
-@UtilityClass
-public class RoomNumberGenerateUtils {
-    private static int roomNumber = 0;
-
-    public int getNextRoomNumber() {
-        return ++roomNumber;
-    }
-
-}
